esp_modem: Move common C definitions in to separate header
Separate header for C API created.
It can be used by C DTE extensions.
