+++ components/esp_modem/idf_component.yml
@@ -1,4 +1,4 @@
-version: "0.1.22"
+version: "0.1.23"
 description: esp modem
 url: https://github.com/espressif/esp-protocols/tree/master/components/esp_modem
 dependencies:
